pyproject.toml
[project]
name = "isl-synthetic-ocr"
version = "0.1.0"
description = "Icelandic OCR Synthetic Data Generation and Training"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "omegaconf>=2.3.0",
    "datasets>=4.1.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "accelerate>=1.10.1",
    "einops>=0.8.1",
    "peft>=0.17.1",
    "wandb>=0.22.0",
    "gradio>=5.47.0",
    "dataclasses>=0.8",
    "ipywidgets>=8.1.7",
    "bitsandbytes>=0.48.1",
    "python-levenshtein>=0.27.3",
    "pyfonts>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_icelandic"]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
